+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 1.3.0
+
+- fixed crash when trying to upload VRChat avatar (you now have to merge animators in Edit mode)
+- added "populate parameters" button
+
 # 1.2.0
 
 - added "debug hide body" flag